refactor(mobile): split mutation registry
